+++ src/partials/modal-window-index-page.html
@@ -9,63 +9,65 @@
   <body> -->
 <button type="button" data-action="open-modal">for test open modal</button>
 <div class="backdrop js-backdrop">
-   <div class="modal_window">
-      <button type="button" class="modal_btn_close" id="btn_mdl_cls" data-action="close-modal">
-         <svg class="modal_icon_close" width="14" height="14" aria-label="icon close">
-            <use href="./images/sprite.svg#icon-close"></use>
-         </svg>
-      </button>
-      <div class="modal_without_close-btn">
-         <img src="./images/replace.jpg" alt="Poster" class="modal_img" />
-         <div class="modal_description">
-            <h1 class="modal_title">Довга назва фільма</h1>
-            <table class="modal_tbl">
-               <tr>
-                  <td class="modal_tbl_left-side" width="50%">Vote / Votes</td>
-                  <td class="modal_tbl_right_top vote" width="10%">0.0</td>
-                  <td class="modal_tbl_right_top" width="40%">/0000</td>
-               </tr>
-               <tr>
-                  <td class="modal_tbl_left-side">Popularity</td>
-                  <td class="modal_tbl_right_top" colspan="2">000.0</td>
-               </tr>
-               <tr>
-                  <td class="modal_tbl_left-side">Original Title</td>
-                  <td class="modal_tbl_right_title" colspan="2">назва</td>
-               </tr>
-               <tr>
-                  <td class="modal_tbl_left-side">Genres</td>
-                  <td class="modal_tbl_right_genre" colspan="2">жанр, жанр, жанр, жанр</td>
-               </tr>
-            </table>
-            <h2 class="modal_title_text">ABOUT</h2>
-            <p class="modal_text">
-               UNFORTUNATELY NO INFORMATION ABOUT THE FILM Lorem ipsum dolor sit amet, Lorem ipsum
-               dolor sit amet consectetur adipisicing elit. Natus cupiditate, officiis ullam tempora
-               commodi blanditiis velit aut animi, aliquam veritatis quos repellat ratione impedit,
-               repudiandae explicabo alias. Debitis, nobis esse. consectetur adipisicing elit. Sed
-               incidunt accusantium, maxime culpa sapiente aliquid consectetur doloribus, sint
-               perferendis, debitis dignissimos? Enim illum delectus magni eligendi reiciendis
-               facilis iure facere. Lorem ipsum dolor sit amet consectetur adipisicing elit. Nostrum
-               distinctio, autem nesciunt aliquid repudiandae a minus vitae eos assumenda quam
-               consequuntur iste delectus eaque alias expedita. Tenetur repudiandae numquam
-               pariatur.
-            </p>
-            <ul class="btn_list">
-               <li>
-                  <button type="button" class="btn_choice btn_watched">
-                     <span class="change_watch">add to Watched</span>
-                  </button>
-               </li>
-               <li>
-                  <button type="button" class="btn_choice btn_queue">
-                     <span class="change_queue">add to queue</span>
-                  </button>
-               </li>
-            </ul>
-         </div>
+  <div class="modal_window">
+    <button
+      type="button"
+      class="modal_btn_close"
+      id="btn_mdl_cls"
+      data-action="close-modal"
+    >
+      <svg
+        class="modal_icon_close"
+        width="14"
+        height="14"
+        aria-label="icon close"
+      >
+        <use href="../images/sprite.svg#icon-close"></use>
+      </svg>
+    </button>
+    <div class="modal_without_close-btn">
+      <img src="./images/replace.jpg" alt="Poster" class="modal_img" />
+      <div class="modal_description">
+        <h1 class="modal_title">Довга назва фільма</h1>
+        <table class="modal_tbl">
+          <tr>
+            <td class="modal_tbl_left-side" width="50%">Vote / Votes</td>
+            <td class="modal_tbl_right_top vote" width="10%">0.0</td>
+            <td class="modal_tbl_right_top" width="40%">/0000</td>
+          </tr>
+          <tr>
+            <td class="modal_tbl_left-side">Popularity</td>
+            <td class="modal_tbl_right_top" colspan="2">000.0</td>
+          </tr>
+          <tr>
+            <td class="modal_tbl_left-side">Original Title</td>
+            <td class="modal_tbl_right_title" colspan="2">назва</td>
+          </tr>
+          <tr>
+            <td class="modal_tbl_left-side">Genres</td>
+            <td class="modal_tbl_right_genre" colspan="2">
+              жанр, жанр, жанр, жанр
+            </td>
+          </tr>
+        </table>
+        <h2 class="modal_title_text">ABOUT</h2>
+        <p class="modal_text">UNFORTUNATELY NO INFORMATION ABOUT THE FILM</p>
+
+        <ul class="btn_list">
+          <li>
+            <button type="button" class="btn_choice btn_watched">
+              <span class="change_watch">add to Watched</span>
+            </button>
+          </li>
+          <li>
+            <button type="button" class="btn_choice btn_queue">
+              <span class="change_queue">add to queue</span>
+            </button>
+          </li>
+        </ul>
       </div>
-   </div>
+    </div>
+  </div>
 </div>
 <!-- <script type="module" src="./index.js"></script>
 </body>
